move Codex catalog discovery into config helpers
