+++ src/ucode/codex_config.py
@@ -1,4 +1,4 @@
-"""Shared helpers for passing Codex configuration on the command line."""
+"""Shared Codex configuration helpers."""
 
 from __future__ import annotations
 
@@ -9,7 +9,9 @@ from pathlib import Path
 import tomlkit
 from tomlkit.items import Item
 
+from ucode.config_io import read_json_safe, read_toml_safe
 from ucode.managed_files import OS, current_os
+from ucode.ui import print_warning
 
 CODEX_PROFILE_NAME = "ucode"
 DEFAULT_CODEX_CONFIG_PATH = Path.home() / ".codex" / f"{CODEX_PROFILE_NAME}.config.toml"
@@ -41,6 +43,52 @@ def codex_config_precedence_paths(
         for path in (managed_config, cli_config, default_config)
         if path is not None
     )
+
+
+def custom_catalog_models() -> list[str] | None:
+    """Read model slugs from the configured model_catalog_json, if present."""
+    try:
+        paths = codex_config_precedence_paths(
+            codex_managed_config_path(),
+            DEFAULT_CODEX_CONFIG_PATH,
+        )
+    except OSError:
+        return None
+    for path in paths:
+        if not path.is_file():
+            continue
+        settings = read_toml_safe(path)
+        catalog_ref = settings.get("model_catalog_json")
+        if not isinstance(catalog_ref, str) or not catalog_ref.strip():
+            continue
+        slugs = _catalog_slugs(Path(catalog_ref).expanduser())
+        if slugs:
+            return slugs
+        print_warning(
+            f"Codex smart routing could not read models from the custom catalog {catalog_ref} "
+            f"referenced by {path}; falling back to the cached model services."
+        )
+        return None
+    return None
+
+
+def _catalog_slugs(path: Path) -> list[str]:
+    """Extract deduplicated model slugs from a Codex custom catalog JSON file."""
+    catalog = read_json_safe(path)
+    models = catalog.get("models")
+    if not isinstance(models, list):
+        return []
+    slugs: list[str] = []
+    seen: set[str] = set()
+    for row in models:
+        if not isinstance(row, dict) or not isinstance(row.get("slug"), str):
+            continue
+        slug = row["slug"].strip()
+        if not slug or slug in seen:
+            continue
+        seen.add(slug)
+        slugs.append(slug)
+    return slugs
 
 
 def _toml_item(value: object) -> Item:
